+++ src/core/injections/controller/checks.py
@@ -55,6 +55,37 @@ else:
     except ImportError:
       settings.READLINE_ERROR = True
 
+
+"""
+check for not declared cookie(s)
+"""
+def not_declared_cookies(response):
+  try:
+    candidate = re.search(r'([^;]+);?', response.headers['set-cookie']).group(1)
+    while True:
+      if not menu.options.batch:
+        question_msg = "You have not declared cookie(s), while "
+        question_msg += "server wants to set its own ('" + str(candidate) + "'). "
+        question_msg += "Do you want to use those [Y/n] > "
+        set_cookies = _input(settings.print_question_msg(question_msg)).lower()
+      else:
+        set_cookies = ""
+      if len(set_cookies) == 0:
+        set_cookies = "Y"
+      if set_cookies in settings.CHOICE_YES:
+        menu.options.cookie = candidate
+        break
+      elif set_cookies in settings.CHOICE_NO: 
+        break
+      elif set_cookies in settings.CHOICE_QUIT:
+        raise SystemExit()
+      else:
+        err_msg = "'" + set_cookies + "' is not a valid answer."  
+        print(settings.print_error_msg(err_msg))
+        pass
+  except KeyError:
+    pass
+
 """
 Tab Autocompleter
 """